requirements.txt
python-docx==1.1.2
pandas==2.2.3
streamlit==1.44.1
openpyxl==3.1.2
xlsxwriter==3.1.2
mistralai==1.5.1

